뉴스레터/requirements.txt
streamlit
requests
beautifulsoup4
jinja2
gspread
google-auth
apscheduler
python-dotenv
openai
